+++ src/Console/Console.php
@@ -13,6 +13,8 @@
 
 namespace Cherry\Console;
 
+use Cherry\Console\Command\Debugger;
+use Cherry\Console\Command\Server;
 use Cherry\Console\Input\ArgvInput;
 use Cherry\Console\Output\Output;
 
@@ -27,6 +29,9 @@ use Cherry\Console\Output\Output;
  */
 class Console
 {
+    use Debugger;
+    use Server;
+
     private $_argvInput;
     private $_output;
 
@@ -82,7 +87,7 @@ class Console
         $method = "_{$method}";
 
         if (method_exists($this, $method)) {
-            $this->{$method}();
+            $this->{$method}($this->_argvInput, $this->_output);
         } else {
             $this->_printHelp();
         }
@@ -95,66 +100,9 @@ class Console
      */
     private function _printHelp()
     {
-        $hello = <<<EOF
-Welcome to
-  ____ _                          
- / ___| |__   ___ _ __ _ __ _   _ 
-| |   | '_ \ / _ \ '__| '__| | | |
-| |___| | | |  __/ |  | |  | |_| |
- \____|_| |_|\___|_|  |_|   \__, |
-                            |___/
-                            
-                          Console
-----------------------------------
-
-help, --help, -h            - Show help.
-    
-server [option] [arguments] - Start PHP Development server.
-       [option]   - Server start options
-            run     - Start server on 127.0.0.1:8000
-            start   - Start server on given address:port
-       [argument] - Additional arguments for option "start" address:port (127.0.0.1:8000)
-            
-EOF;
+        $help = file_get_contents(__DIR__.'/Files/Docs/help.txt');
 
         print $this->_output
-            ->text($hello);
-    }
-
-    /**
-     * Run PHP Development server from CLI
-     *
-     * @return void
-     */
-    private function _server()
-    {
-        $argv = $this->_argv;
-
-        $server = null;
-
-        if (isset($argv[1]) && $argv[1] == 'run') {
-            $server = "127.0.0.1:8000";
-        } else if (isset($argv[1]) && $argv[1] == 'start') {
-            if (isset($argv[2])) {
-                $server = $argv[2];
-            } else {
-                $this->_printHelp();
-                return;
-            }
-        } else {
-            $this->_printHelp();
-            return;
-        }
-
-        $info = <<<EOF
-        
-{$this->_output->success("Started Cherry Server on  http://{$server}")}
-// Quit the server with Ctrl + C.
-EOF;
-
-        print $this->_output
-            ->text($info);
-
-        echo exec("php -S {$server} -t " . WEB_ROOT);
+            ->text($help);
     }
 }